+++ R/transition-fnc.R
@@ -27,7 +27,14 @@
 #'
 #' @note
 #'
-#' To create a preprocessing function you can use `dplyr` or `data.table`. You can
+#' In general, dymiumCore detects variables in entity data of an [Entity] object that
+#' has a dot prefix as derived variables. Meaning, those derived variables are not
+#' check against new entity data that are getting added to the [Entity] object. But
+#' when entity data are used in `get_transition`, the dot prefix of the derived variables
+#' will be removed. This is to make it convenient when naming variables during
+#' the model estimation step.
+#'
+#' To create a pre-processing function you can use `dplyr` or `data.table`. You can
 #' even combine multiple functions with [magrittr::%>%]. As an example, if you only
 #' want to filter just the male population then you can choose one of the
 #' following options to create your preprocessing function.
@@ -237,6 +244,7 @@ get_transition <- function(world, entity, model, target = NULL, targeted_ids = N
     e_data <-
       model$preprocessing_fn(e_data)
   }
+  e_data <- dymiumCore::normalise_derived_vars(e_data)
   if (nrow(e_data) == 0) {
     return(data.table(id = integer(), response = character()))
   }