+++ NEWS.md
@@ -3,6 +3,7 @@
 - Added `sim()` for compiling and executing a microsimulation pipeline.
 - Added `targets` as an active field. 
 - Fixed `get_transition` didn't use the preprocessing function when the `model` argument is of `Model` class.
+- `get_transition` now removes the dot prefix of variables in entity data. Those variables are considered as derived variables which will not be checked against when new entity data are geting added.
 
 # dymiumCore 0.1.7